requirements.txt
torch~=1.8.1
pillow~=8.2.0
pytorch-ignite~=0.4.4
numpy~=1.19.2
requests~=2.25.1
PyMySQL~=1.0.2
torchvision~=0.9.1
torchsummary~=1.5.1
matplotlib